backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
python-dotenv
elevenlabs